OCR/requirements.txt

fastapi
Pillow
gunicorn
uvicorn
numpy
python-multipart
easyocr